GoogleAnalytics: Fallback
If 'analytics.js' fails to load from google-analytics.com, fallback
to a local file so the application doesn't crash.

If AJAX request fails because it's blocked, this will not have any
effect (except error messages in dev-tools in browser). For most
users it will silently fail, which is okay - though it hurts
statistics.

Solves: Issue 4